+++ src/starmatrix/model.py
@@ -1,13 +1,13 @@
 import math
 import numpy as np
-import intergalactic.constants as constants
-import intergalactic.elements as elements
-import intergalactic.matrix as matrix
-from intergalactic.imfs import select_imf
-from intergalactic.abundances import select_abundances
-from intergalactic.dtds import select_dtd, dtd_correction
-from intergalactic.functions import stellar_mass, stellar_lifetime, max_mass_allowed, return_fraction
-from intergalactic.functions import total_energy_ejected, newton_cotes, global_imf, imf_supernovas_II
+import starmatrix.constants as constants
+import starmatrix.elements as elements
+import starmatrix.matrix as matrix
+from starmatrix.imfs import select_imf
+from starmatrix.abundances import select_abundances
+from starmatrix.dtds import select_dtd, dtd_correction
+from starmatrix.functions import stellar_mass, stellar_lifetime, max_mass_allowed, return_fraction
+from starmatrix.functions import total_energy_ejected, newton_cotes, global_imf, imf_supernovas_II
 
 
 class Model: