+++ src/starmatrix/__init__.py
@@ -5,4 +5,4 @@ Starmatrix
 Generation of matrices of elements in spiral galaxies
 
 """
-__version__ = "1.2.0"
+__version__ = "1.3.0"