+++ CHANGELOG.rst
@@ -5,7 +5,7 @@ Changelog
 =========
 
 
-1.3.0 (Unreleased)
+1.3.0 (2020-12-09)
 ==================
 
 - SuperNova rates are now computed accounting for the total number of stars created per unit of stellar mass